django/requirements.txt
psycopg2
django-mptt
